Change TableValueStyleCache to interface
