+++ atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/impl/TableValueStyleCache.java
@@ -15,50 +15,12 @@
  */
 package com.palantir.atlasdb.keyvalue.dbkvs.impl;
 
-import com.google.common.base.Preconditions;
-import com.google.common.base.Throwables;
-import com.google.common.cache.Cache;
-import com.google.common.cache.CacheBuilder;
-import com.google.common.collect.Iterables;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.logging.LoggingArgs;
-import com.palantir.logsafe.logger.SafeLogger;
-import com.palantir.logsafe.logger.SafeLoggerFactory;
-import com.palantir.nexus.db.sql.AgnosticResultSet;
-import com.palantir.nexus.db.sql.SqlConnection;
-import java.util.concurrent.ExecutionException;
 
-public class TableValueStyleCache {
-    private static final SafeLogger log = SafeLoggerFactory.get(TableValueStyleCache.class);
+public interface TableValueStyleCache {
 
-    private final Cache<TableReference, TableValueStyle> valueStyleByTableRef =
-            CacheBuilder.newBuilder().build();
+    TableValueStyle getTableType(
+            ConnectionSupplier connectionSupplier, TableReference tableRef, TableReference metadataTable);
 
-    public TableValueStyle getTableType(
-            final ConnectionSupplier connectionSupplier, final TableReference tableRef, TableReference metadataTable) {
-        try {
-            return valueStyleByTableRef.get(tableRef, () -> {
-                SqlConnection conn = connectionSupplier.get();
-                AgnosticResultSet results = conn.selectResultSetUnregisteredQuery(
-                        String.format(
-                                "SELECT table_size FROM %s WHERE table_name = ?", metadataTable.getQualifiedName()),
-                        tableRef.getQualifiedName());
-                Preconditions.checkArgument(
-                        !results.rows().isEmpty(), "table %s not found", tableRef.getQualifiedName());
-
-                return TableValueStyle.byId(
-                        Iterables.getOnlyElement(results.rows()).getInteger("table_size"));
-            });
-        } catch (ExecutionException e) {
-            log.error(
-                    "TableValueStyle for the table {} could not be retrieved.",
-                    LoggingArgs.safeInternalTableName(tableRef.getQualifiedName()),
-                    e);
-            throw Throwables.propagate(e);
-        }
-    }
-
-    public void clearCacheForTable(TableReference tableRef) {
-        valueStyleByTableRef.invalidate(tableRef);
-    }
+    void clearCacheForTable(TableReference tableRef);
 }
